added error handling

+++ lib/screens/welcome/signup/signup_screen.dart
@@ -2,6 +2,7 @@ import 'dart:developer';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import '../../components/custom_snack_bar_content.dart';
 import '../../topic/topic_screen.dart';
 import '../components/already_have_an_account_check.dart';
 import '../components/rounded_button.dart';
@@ -32,37 +33,63 @@ class _SignUpScreenState extends State<SignUpScreen> {
     String password = _passwordController.text.trim();
 
     try {
-      UserCredential userCredential = await FirebaseAuth.instance
-          .createUserWithEmailAndPassword(email: email, password: password);
-      if (userCredential.user != null) {
-        Map<String, dynamic> userData = {
-          "Name": name,
-          "Email": email,
-          "Score": 0,
-          "Phone Number": "000 000000",
-          "Date of Birth": "Select",
-          "Profile Pic":
-              "https://firebasestorage.googleapis.com/v0/b/quizapp-x-flutter.appspot.com/o/defaultProfilePic%2FdefaultProfilePic.png?alt=media&token=c65b9636-04f4-4a19-aa19-67a9629a6ee5",
-        };
-        try {
-          await _firestore.collection("users").doc(email).set(userData);
-          if (context.mounted) {
-            Navigator.popUntil(context, (route) => route.isFirst);
-            Navigator.pushReplacementNamed(context, TopicScreen.id);
+      if (name != "" && email != "" && password != "") {
+        UserCredential userCredential = await FirebaseAuth.instance
+            .createUserWithEmailAndPassword(email: email, password: password);
+        if (userCredential.user != null) {
+          Map<String, dynamic> userData = {
+            "Name": name,
+            "Email": email,
+            "Score": 0,
+            "Phone Number": "000 000000",
+            "Date of Birth": "Select",
+            "Profile Pic":
+                "https://firebasestorage.googleapis.com/v0/b/quizapp-x-flutter.appspot.com/o/defaultProfilePic%2FdefaultProfilePic.png?alt=media&token=c65b9636-04f4-4a19-aa19-67a9629a6ee5",
+          };
+          try {
+            await _firestore.collection("users").doc(email).set(userData);
+            if (context.mounted) {
+              Navigator.popUntil(context, (route) => route.isFirst);
+              Navigator.pushReplacementNamed(context, TopicScreen.id);
+            }
+            setState(() {
+              _emailController.clear();
+              _passwordController.clear();
+              _nameController.clear();
+            });
+          } on FirebaseException catch (e) {
+            userCredential.user!.delete();
+            log("Second Error: ${e.code}");
           }
-          setState(() {
-            _emailController.clear();
-            _passwordController.clear();
-            _nameController.clear();
-          });
-        } on FirebaseException catch (e) {
-          userCredential.user!.delete();
-          log(e.code.toString());
         }
+      } else {
+        ScaffoldMessenger.of(context).showSnackBar(
+          const SnackBar(
+            backgroundColor: Colors.transparent,
+            elevation: 0,
+            content: CustomSnackBarContent(
+              error: 'Registration Error',
+              explanation: "All the field are required",
+            ),
+            behavior: SnackBarBehavior.floating,
+            duration: Duration(seconds: 3),
+          ),
+        );
       }
     } on FirebaseAuthException catch (e) {
-      //TODO: show snack-bar
-      log(e.code.toString());
+      log("First Error: ${e.code}");
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          backgroundColor: Colors.transparent,
+          elevation: 0,
+          content: CustomSnackBarContent(
+            error: 'Registration Error',
+            explanation: e.code.toString(),
+          ),
+          behavior: SnackBarBehavior.floating,
+          duration: const Duration(seconds: 3),
+        ),
+      );
     }
   }
